+++ server/src/main/java/com/example/demo/mapper/TagsMapper.java
@@ -7,5 +7,5 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface TagsMapper {
-  List<Tags> getTagsByIds(@Param("ids") List<Long> ids);
+  List<Tags> getTagsByIds(@Param("ids") List<Integer> ids);
 }
